requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
groq==0.9.0
PyMuPDF==1.24.9
